refactor: readme & login account not verified
