+++ README.md
@@ -1,30 +1,103 @@
-# React + TypeScript + Vite
+# Football Field Booking Website
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+## Target
 
-Currently, two official plugins are available:
+- Create website where user can come and book football field.
+- The primary goal is to provide a seamless experience for users to easily find and book football fields in their area.
+- Enable users to search for available fields based on location, date, time, and other preferences.
+- Ensure responsiveness across various devices to cater to users accessing the platform from desktops, laptops, tablets, and mobile phones.
+- Prioritize security and privacy measures to safeguard user information and payment transactions.
+- Offer features such as user accounts, booking history, and notifications to enhance the overall user experience and encourage return visits.
+- Collaborate with field owners or administrators to streamline the process of adding and managing field listings on the platform.
+- Incorporate feedback mechanisms to gather user insights and continuously improve the platform based on user preferences and suggestions.
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+## Design
 
-## Expanding the ESLint configuration
+[Figma](https://www.figma.com/file/ihotGtX1kFp8bnKwGMzyQl/Final-Project?type=design&node-id=0%3A1&mode=design&t=Krsnt0yxeNoZzcNK-1)
 
-If you are developing a production application, we recommend updating the configuration to enable type aware lint rules:
+## Information
 
-- Configure the top-level `parserOptions` property like this:
+- Editor: Visual Studio Code
+- Supported browser: Chrome, Firefox, MS Edge, Opera, Safari lasted
+- Supported screen: Screen width 996px or larger
 
-```js
-export default {
-  // other rules...
-  parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module',
-    project: ['./tsconfig.json', './tsconfig.node.json'],
-    tsconfigRootDir: __dirname,
-  },
-}
+## Team size
+
+- Dev: Liem Nguyen
+- Dev: Le Hoang
+
+## Develop Environment
+
+- [Visual Studio Code](https://code.visualstudio.com/)
+- [Github](https://github.com/)
+
+## Technical used in this project
+
+- React
+- TypeScript
+- Vite
+- React Router Dom
+- TailwindCSS
+- Lucide React _(Icon library)_
+- Shadcn/UI _(Radix-UI)_
+- GSAP _(Animation Library)_
+- Tanstack
+- Zustand
+- Zod _(Validating)_
+- MapboxGl _(Map API)_
+
+## Folder structure
+
+```bash
+.
+├── src/
+│   ├── assets/
+│   ├── components/
+│   ├── constants/
+│   ├── hooks/
+│   ├── pages/
+│   │   ├── Login/
+│   │       ├── components/
+│   │       ├── index.tsx
+│   ├── routes/
+│   ├── services/
+│   ├── stores/
+│   ├── types/
+│   ├── utils/
+│   └── App.tsx
+│   └── main.tsx
+│   └── index.css
+└── README.md
 ```
 
-- Replace `plugin:@typescript-eslint/recommended` to `plugin:@typescript-eslint/recommended-type-checked` or `plugin:@typescript-eslint/strict-type-checked`
-- Optionally add `plugin:@typescript-eslint/stylistic-type-checked`
-- Install [eslint-plugin-react](https://github.com/jsx-eslint/eslint-plugin-react) and add `plugin:react/recommended` & `plugin:react/jsx-runtime` to the `extends` list
+## Getting started
+
+- Step 01: Clone repository with HTTPS
+
+```bash
+git clone https://github.com/liemgumball/football_field_booking-FE.git
+```
+
+- Step 02: Move to folder which just cloned in your computer
+
+```bash
+cd football_field_booking-FE
+```
+
+- Step 03: Install packages
+
+```bash
+yarn
+```
+
+- Step 04: Build
+
+```bash
+yarn build
+```
+
+- Step 05: Run
+
+```bash
+yarn preview
+```
