+++ src/constants/message.ts
@@ -1,5 +1,6 @@
 export const ERROR_MSG = {
 	ERROR_FALLBACK: 'An error has occurred',
 	LOGIN_FAILED: 'Wrong email or password',
+	ACCOUNT_NOT_VERIFIED: 'Your account is not verified',
 	SERVER_ERROR: 'Internal server error',
 } as const
